+++ includes/ServiceModules/MybbExtraGroups/ServiceMybbExtraGroupsSimple.php
@@ -9,7 +9,7 @@ use App\View\Html\Wrapper;
 use App\ServiceModules\Interfaces\IServiceAdminManage;
 use App\ServiceModules\Interfaces\IServiceCreate;
 use App\ServiceModules\Interfaces\IServiceUserServiceAdminDisplay;
-use App\ServiceModules\Service;
+use App\ServiceModules\ServiceModule;
 use App\View\CurrentPage;
 use App\System\Settings;
 use App\Translation\TranslationManager;
@@ -17,7 +17,7 @@ use App\Translation\Translator;
 
 // TODO Remove Simple classes. Rename Service to ServiceModule
 
-abstract class ServiceMybbExtraGroupsSimple extends Service implements
+abstract class ServiceMybbExtraGroupsSimple extends ServiceModule implements
     IServiceAdminManage,
     IServiceCreate,
     IServiceUserServiceAdminDisplay