+++ CHANGES.md
@@ -8,3 +8,4 @@
 * Moved App/Pages
 * Moved App/System/CurrentPage
 * Moved App/Services to App/ServiceModules
+* class Service to ServiceModule